% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_slack_api.R
\name{auth_test}
\alias{auth_test}
\title{Checks authentication & identity against the Slack API.}
\usage{
auth_test(bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN"))
}
\arguments{
\item{bot_user_oauth_token}{The Slack bot OAuth token {character vector}}
}
\description{
Checks authentication & identity against the Slack API.
}
\examples{
if (Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN") != "") {
  auth_test()
}
}
\references{
https://api.slack.com/methods/auth.test
}
