% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_slackr.R
\name{check_tex_pkg}
\alias{check_tex_pkg}
\title{check_tex_pkg}
\usage{
check_tex_pkg()
}
\description{
Install or load texPreview package,
  inspired by the \code{parsnip} package
}
\details{
Check if texPreview is intalled
}
