% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{check_tokens}
\alias{check_tokens}
\title{Check for token issues}
\usage{
check_tokens(token, bot_user_oauth_token)
}
\arguments{
\item{token}{a token}

\item{bot_user_oauth_token}{another token}
}
\value{
A token
}
\description{
Check for token issues
}
