% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checknew}
\alias{checknew}
\title{Checks a specific slack channel for updated messages.}
\usage{
checknew(oldest = as.numeric(Sys.time()),
  channel = Sys.getenv("SLACK_CHANNEL"),
  username = Sys.getenv("SLACK_USERNAME"),
  icon_emoji = Sys.getenv("SLACK_ICON_EMOJI"),
  api_token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{oldest}{the datetime (as numeric) from which to start looking at messages.}

\item{channel}{slack channel}

\item{username}{slack username to respond as}

\item{icon_emoji}{slack emoji to use}

\item{api_token}{slack api token}
}
\description{
Needs some updates here.
}

