% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkslack}
\alias{checkslack}
\title{main function, will loop and watch for messages}
\usage{
checkslack(latest = as.numeric(Sys.time()), sleep = 2,
  channel = Sys.getenv("SLACK_CHANNEL"),
  username = Sys.getenv("SLACK_USERNAME"),
  icon_emoji = Sys.getenv("SLACK_ICON_EMOJI"),
  api_token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{latest}{the datetime (as numeric) from which to start looking at messages.}

\item{sleep}{the amount of time between checks in seconds}

\item{channel}{slack channel}

\item{username}{slack username to respond as}

\item{icon_emoji}{slack emoji to use}

\item{api_token}{slack api token}
}
\description{
Needs some updates here.
}

