% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_slack_api.R
\name{convert_response_to_tibble}
\alias{convert_response_to_tibble}
\title{Convert Slack API json response to tibble.}
\usage{
convert_response_to_tibble(x, element)
}
\arguments{
\item{x}{The Slack API response object, returned from \link{call_slack_api}}

\item{element}{The name of the list element to extract}
}
\value{
A tibble
}
\description{
Convert Slack API json response to tibble.
}
\keyword{Internal}
