% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_slackr.R
\name{delete_slackr}
\alias{delete_slackr}
\title{delete messages of a slack channel. Calls the chat.delete method on the Slack Web API.
Information on this method can be found here: \url{https://api.slack.com/methods/chat.delete}}
\usage{
delete_slackr(
  count,
  channel = Sys.getenv("SLACK_CHANNEL"),
  api_token = Sys.getenv("SLACK_API_TOKEN")
)
}
\arguments{
\item{count}{Number of messages to delete.}

\item{channel}{The name of the channels to delete messages from}

\item{api_token}{your full Slack API token}
}
\value{
\code{httr} response object (invislbly)
}
\description{
delete messages of a slack channel. Calls the chat.delete method on the Slack Web API.
Information on this method can be found here: \url{https://api.slack.com/methods/chat.delete}
}
\examples{
\dontrun{
slackr_setup()
delete_slackr(count=1)
}
}
\seealso{
\url{https://api.slack.com/methods/chat.delete}
}
\author{
Quinn Weber [aut], Bob Rudis [ctb]

Jonathan Sidi [aut]
}
