% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcamels.r
\name{dev.slackr}
\alias{dev.slackr}
\title{dev.slackr}
\usage{
dev.slackr(...)
}
\arguments{
\item{...}{arguments to pass to slackr_dev}
}
\description{
deprecated
}
\seealso{
\code{\link[slackr]{slackr_dev}}
}
