% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_slackr.R
\name{edit_slackr}
\alias{edit_slackr}
\title{edit a message on a slack channel. Calls the chat.update method on the Slack Web API.
Information on this method can be found here: \url{https://api.slack.com/methods/chat.update}}
\usage{
edit_slackr(
  text,
  pattern = NULL,
  idx = 1,
  hs = NULL,
  ...,
  preformatted = TRUE,
  channel = Sys.getenv("SLACK_CHANNEL"),
  username = Sys.getenv("SLACK_USERNAME"),
  icon_emoji = Sys.getenv("SLACK_ICON_EMOJI"),
  api_token = Sys.getenv("SLACK_API_TOKEN")
)
}
\arguments{
\item{text}{The character vector to be posted}

\item{pattern}{Filter messages by regex (grepl), Default: NULL}

\item{idx}{Index of message to edit (descending order), Default: 1}

\item{hs}{History of slack channel (if NULL a request to slacka api will be made), Default=NULL}

\item{...}{Optional arguments such as: as_user, parse, unfurl_links, etc.}

\item{preformatted}{Should the text be sent as preformatted text. Defaults to TRUE}

\item{channel}{The name of the channels to which the DataTable should be sent.
Prepend channel names with a hashtag. Prepend private-groups with nothing.
Prepend direct messages with an @}

\item{username}{what user should the bot be named as (chr)}

\item{icon_emoji}{what emoji to use (chr) \code{""} will mean use the default}

\item{api_token}{your full Slack API token}
}
\value{
\code{httr} response object (invislbly)
}
\description{
edit a message on a slack channel. Calls the chat.update method on the Slack Web API.
Information on this method can be found here: \url{https://api.slack.com/methods/chat.update}
}
\note{
You can pass in \code{add_user=TRUE} as part of the \code{...} parameters and the Slack API
      will post the message as your logged-in user account (this will override anything set in
      \code{username})
}
\examples{
\dontrun{
slackr_setup()

text_slackr('hello world')
text_slackr('hello new world')

edit_slackr('another new world')
edit_slackr('goodbye new world',pattern='world',idx=2)

}
}
\seealso{
\url{https://api.slack.com/methods/chat.update}
}
\author{
Jonathan Sidi [aut]
}
