% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_slackr.R, R/slackr_utils.R
\name{history_slackr}
\alias{history_slackr}
\alias{slackr_channel_history}
\title{return message history of a slack channel. Calls the channels.history method on the Slack Web API.
Information on this method can be found here: \url{https://api.slack.com/methods/channels.history}}
\usage{
history_slackr(count, ..., channel = Sys.getenv("SLACK_CHANNEL"),
  api_token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{count}{Number of messages to return, between 1 and 1000.}

\item{...}{Optional arguments such as: inclusive, latest, oldest, unreads}

\item{channel}{The name of the channels to which the DataTable should be sent.
Prepend channel names with a hashtag. Prepend private-groups with nothing.
Prepend direct messages with an @}

\item{api_token}{your full Slack API token}
}
\value{
\code{httr} response object (invislbly)
}
\description{
return message history of a slack channel. Calls the channels.history method on the Slack Web API.
Information on this method can be found here: \url{https://api.slack.com/methods/channels.history}
}
\examples{
\dontrun{
slackr_setup()
history_slackr(count=1)
}
}
\seealso{
\url{https://api.slack.com/methods/channels.history}
}
\author{
Jonathan Sidi [aut]
}
