% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_slackr.R
\name{htmlslackr}
\alias{htmlslackr}
\title{Post an htmlwidget to a Slack channel}
\usage{
htmlslackr(plot, title = "widget", initial_comment = NULL,
  channels = Sys.getenv("SLACK_CHANNEL"),
  api_token = Sys.getenv("SLACK_API_TOKEN"), file = "widget", ...)
}
\arguments{
\item{plot}{htmlwidget object to save, defaults to last item ouptut}

\item{title}{title on Slack (optional - defaults to filename)}

\item{initial_comment}{comment for file on slack (optional - defaults to filename)}

\item{channels}{list of channels to post image to}

\item{api_token}{the Slack full API token (chr)}

\item{file}{prefix for filenames (defaults to \code{widget})}

\item{...}{other arguments passed to graphics device}
}
\value{
\code{httr} response object (invisibly)
}
\description{
Unlike the \code{\link{dev_slackr}} function, this one takes an \code{htmlwidget} object,
allowing the user to post a static image of an htmlwidget to a slack channel.
}
\note{
You need to setup a full API token (i.e. not a webhook & not OAuth) for this to work
      Also, uou can pass in \code{as_user=TRUE} as part of the \code{...}
      parameters and the Slack API will post the message as your logged-in user
      account (this will override anything set in \code{username})
}
\examples{
\dontrun{
slackr_setup()
htmlslackr(DT::datatable(iris))
}
}
