% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_slack_internals.R
\name{list_channels}
\alias{list_channels}
\title{Lists all channels in a Slack team.}
\usage{
list_channels(
  bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN"),
  types = "public_channel",
  ...
)
}
\description{
Lists all channels in a Slack team.
}
\references{
https://api.slack.com/methods/conversations.list
}
\keyword{internal}
