% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lookupUser}
\alias{lookupUser}
\title{given username, returns the api ID}
\usage{
lookupUser(user, api_token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{user}{the slack username (must be preceded by "at" symbol)}

\item{api_token}{slack api token}
}
\description{
Needs some updates here.
}

