% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_slack_internals.R
\name{post_message}
\alias{post_message}
\title{Sends a message to a channel.}
\usage{
post_message(
  txt,
  channel,
  bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN"),
  ...
)
}
\description{
Sends a message to a channel.
}
\references{
https://api.slack.com/methods/chat.postMessage
}
