% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_onexit.R
\name{register_onexit}
\alias{register_onexit}
\title{Append text_slackr as on.exit to functions.}
\usage{
register_onexit(f, msg = "done!", env = parent.frame(2),
  channel = Sys.getenv("SLACK_CHANNEL"),
  username = Sys.getenv("SLACK_USERNAME"),
  icon_emoji = Sys.getenv("SLACK_ICON_EMOJI"),
  api_token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{f}{function or character}

\item{msg}{character, message to pass to text_slackr Default: 'done!'}

\item{env}{environment to assign appended function to with relation to the function environment,
Default: parent.frame(2) (global environment)}

\item{channel}{The name of the channels to which the DataTable should be sent.
Prepend channel names with a hashtag. Prepend private-groups with nothing.
Prepend direct messages with an @}

\item{username}{what user should the bot be named as (chr)}

\item{icon_emoji}{what emoji to use (chr) \code{""} will mean use the default}

\item{api_token}{your full Slack API token}
}
\value{
function
}
\description{
Appends to the body of a function an on.exit call to run at the end of the call.
}
\details{
If a character is passed to f then it will evaluate internally to a function.
}
\examples{
\dontrun{
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
group <- gl(2, 10, 20, labels = c("Ctl","Trt"))
weight <- c(ctl, trt)
register_onexit(lm,channel="general")
lm.D9 <- slack_lm(weight ~ group)
}
}
\references{
\url{https://github.com/hrbrmstr/slackr/pull/11}
}
\seealso{
\code{\link{text_slackr}}
}
\author{
Jonathan Sidi [aut]
}
