% Generated by roxygen2 (4.0.1): do not edit by hand
\name{respond}
\alias{respond}
\title{Figure out what response is necessary and send it.}
\usage{
respond(msg, channel = Sys.getenv("SLACK_CHANNEL"),
  username = Sys.getenv("SLACK_USERNAME"),
  icon_emoji = Sys.getenv("SLACK_ICON_EMOJI"),
  api_token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{msg}{the slack message (as a list)}

\item{channel}{slack channel}

\item{username}{slack username to respond as}

\item{icon_emoji}{slack emoji to use}

\item{api_token}{slack api token}
}
\description{
Needs some updates here.
}

