% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runcensus.R
\name{runcensus}
\alias{runcensus}
\title{Get a census of all users, channels, and groups in your Slack workspace}
\usage{
runcensus(api_token)
}
\arguments{
\item{api_token}{the Slack full API token (chr)}
}
\value{
list consisting of
\itemize{
  \item{channels}
  \item{ims}
  \item{users}
  \item{groups}
}
}
\description{
Gathers all the channels, users, ims and groups of your slack workspace into a single object.
}
\examples{
\dontrun{
creds <- yaml::yaml.load_file(input = "~/src/slackr_creds.yml")
slackr_census <- slackr_runcensus(creds)
}
}
\author{
Daniel Egan [ctb]
}
