% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcamels.r
\name{save.slackr}
\alias{save.slackr}
\title{save.slackr}
\usage{
save.slackr(...)
}
\arguments{
\item{...}{arguments to pass to save_slackr}
}
\description{
deprecated
}
\seealso{
\code{\link[slackr]{save_slackr}}
}
