% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/send_message.R
\name{send_message}
\alias{send_message}
\title{Send a message to slack from R console.}
\usage{
send_message(text, channel = DEFAULT_CHANNEL(), name = "slackr",
  icon_url = DEFAULT_ICON(), attachments = NULL)
}
\description{
Send a message to slack from R console.
}

