% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{package}
\name{slackr-package}
\alias{slackr-package}
\title{slackr-package}
\description{
A package to work with Slack.com webhook API messages
}
\details{
You'll probably want to dive right into the \link{slackr} function for more information.
}
\examples{
\dontrun{
slackrSetup()
slackr("iris info", head(iris), str(iris))
}
}
\author{
Bob Rudis (@hrbrmstr)
}

