% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{package}
\name{slackr}
\alias{slackr}
\alias{slackr-package}
\title{A package to send webhook API messages to Slack.com channels/users}
\usage{
slackr(..., channel = Sys.getenv("SLACK_CHANNEL"),
  username = Sys.getenv("SLACK_USERNAME"),
  icon_emoji = Sys.getenv("SLACK_ICON_EMOJI"),
  slack_webhook_url_prefix = Sys.getenv("SLACK_URL_PREFIX"),
  token = Sys.getenv("SLACK_TOKEN"))
}
\arguments{
\item{...}{expressions to be sent to Slack.com}

\item{channel}{which channel to post the message to (chr)}

\item{username}{what user should the bot be named as (chr)}

\item{icon_emoji}{what emoji to use (chr) \code{""} will mean use the default}

\item{slack_webhook_url_prefix}{which slack.com API endpoint URL to use}

\item{token}{your webhook API token}
}
\description{
A package to send webhook API messages to Slack.com channels/users

Takes an \code{expr}, evaluates it and sends the output to a Slack.com
chat destination. Useful for logging, messaging on long compute tasks or
general information sharing.
}
\details{
By default, everyting but \code{expr} will be looked for in a "\code{SLACK_}"
environment variable. You can override or just specify these values directly instead,
but it's probably better to call \code{slackrSetup()} first.
}
\examples{
\dontrun{
slackrSetup(channel="#code", url_prefix="http://myslack.slack.com/services/hooks/incoming-webhook?")
slackr(str(iris))
}
\dontrun{
slackrSetup(channel="#code",
            url_prefix="http://myslack.slack.com/services/hooks/incoming-webhook?")
slackr(str(iris))
}
}
\author{
Bob Rudis (@hrbrmstr)
}

