% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr_utils.R
\name{slackr_census}
\alias{slackr_census}
\title{Create a cache of the users and channels in the workspace in order to limit API requests}
\usage{
slackr_census(bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN"))
}
\arguments{
\item{bot_user_oauth_token}{the Slack bot OAuth token (chr)}
}
\value{
A data.frame of channels and users
}
\description{
Create a cache of the users and channels in the workspace in order to limit API requests
}
