% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr_utils.R, R/zcamels.r
\name{slackr_chtrans}
\alias{slackr_chtrans}
\alias{slackrChtrans}
\alias{slackrChTrans}
\title{Translate vector of channel names to channel IDs for API}
\usage{
slackr_chtrans(channels)

slackrChtrans(...)

slackrChTrans(...)
}
\arguments{
\item{channels}{vector of channel names to parse}

\item{...}{
  Arguments passed on to \code{\link[=dev.slackr]{dev.slackr}}, \code{\link[=dev.slackr]{dev.slackr}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
character vector - original channel list with \verb{#} or
\code{@} channels replaced with ID's.
}
\description{
Given a vector of one or more channel names, retrieve list of
active channels and try to replace channels that begin with "\verb{#}" or "\code{@}"
with the channel ID for that channel.
}
\author{
Quinn Weber (ctb), Bob Rudis (aut)
}
