% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr_utils.R
\name{slackr_conversations}
\alias{slackr_conversations}
\title{Get a data frame of Slack conversations}
\usage{
slackr_conversations(api_token = Sys.getenv("SLACK_API_TOKEN"),
  types = "public_channel")
}
\arguments{
\item{api_token}{the Slack full API token (chr)}

\item{types}{Types of conversations to pull, comma separated, e.g. public_channel,im will pull public channels and direct messages. The possible values are public_channel, private_channel, im and mpim}
}
\value{
data.table of conversations
}
\description{
Get a data frame of Slack conversations
}
