% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr_utils.R
\name{slackr_createcache}
\alias{slackr_createcache}
\title{Create a cache of the users and channels in the workspace in order to limit API requests}
\usage{
slackr_createcache()
}
\value{
the memoized census function
}
\description{
Create a cache of the users and channels in the workspace in order to limit API requests
}
