% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr_delete.R
\name{slackr_delete}
\alias{slackr_delete}
\title{slackr_delete}
\usage{
slackr_delete(
  count,
  channel = Sys.getenv("SLACK_CHANNEL"),
  token = Sys.getenv("SLACK_TOKEN")
)
}
\arguments{
\item{count}{the number of messages to delete}

\item{channel}{Channel, private group, or IM channel to delete messages from. Can be an encoded ID, or a name. See https://api.slack.com/methods/chat.postMessage#channels for more details.}

\item{token}{Authentication token bearing required scopes. Tokens should be passed as an HTTP Authorization header or alternatively, as a POST parameter.}
}
\description{
Deletes the specified number of messages from the channel
}
