% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr_utils.R, R/zcamels.r
\name{slackr_ims}
\alias{slackr_ims}
\alias{slackrIms}
\title{Get a data frame of Slack IM ids}
\usage{
slackr_ims(
  bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN"),
  verbose = interactive()
)

slackrIms(...)
}
\arguments{
\item{bot_user_oauth_token}{the Slack both OAuth token (chr)}

\item{verbose}{If TRUE, prints progress messages}

\item{...}{
  Arguments passed on to \code{\link[=dev.slackr]{dev.slackr}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\code{data.frame} of im ids and user names
}
\description{
Get a data frame of Slack IM ids
}
\references{
\url{https://github.com/mrkaye97/slackr/pull/13}
}
\author{
Quinn Weber (aut), Bob Rudis (ctb)
}
