% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr_setup.r
\name{slackr_teardown}
\alias{slackr_teardown}
\title{Unset env vars created by \code{slackr_setup()}}
\usage{
slackr_teardown()
}
\value{
TRUE if successful (invisibly)
}
\description{
Unset env vars created by \code{slackr_setup()}
}
\examples{
\dontrun{
  slackr_teardown()
}
}
\seealso{
\code{\link[=slackr_setup]{slackr_setup()}}
}
