% Generated by roxygen2 (4.0.1): do not edit by hand
\name{slackr_users}
\alias{slackr_users}
\title{Get data frame of slack.com users}
\usage{
slackr_users(api_token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{api_token}{the slack.com full API token (chr)}
}
\description{
need to setup a full API token (i.e. not a webhook & not OAuth) for this to work
}

