% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcamels.r
\name{slackrChannels}
\alias{slackrChannels}
\title{slackrChannels}
\usage{
slackrChannels(...)
}
\arguments{
\item{...}{arguments to pass to slackr_channels}
}
\description{
deprecated
}
\seealso{
\code{\link[slackr]{slackr_channels}}
}
