% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcamels.r
\name{slackrChtrans}
\alias{slackrChtrans}
\alias{slackrChTrans}
\title{slackrChtrans}
\usage{
slackrChtrans(...)

slackrChTrans(...)
}
\arguments{
\item{...}{arguments to pass to slackr_chtrans}
}
\description{
deprecated

deprecated
}
\seealso{
\code{\link[slackr]{slackr_chtrans}}

\code{\link[slackr]{slackr_chtrans}}
}
