% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/slackr.R
\name{slackrIms}
\alias{slackrIms}
\title{Get a data frame of slack.com IM ids}
\usage{
slackrIms(api_token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{api_token}{the slack.com full API token (chr)}
}
\value{
data.table of im ids and user names
}
\description{
need to setup a full API token (i.e. not a webhook & not OAuth) for this to work
}

