% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\name{slackrMsg}
\alias{slackrMsg}
\title{Superseded by slackr_msg}
\usage{
slackrMsg(...)
}
\arguments{
\item{...}{arguments to pass to slackr_msg}
}
\description{
deprecated
}
\seealso{
\code{\link[slackr]{slackr_msg}}

Other Deprecated functions: 
\code{\link{dev.slackr}()},
\code{\link{dev_slackr}()},
\code{\link{save.slackr}()},
\code{\link{slackrBot}()},
\code{\link{slackrChannels}()},
\code{\link{slackrChtrans}()},
\code{\link{slackrIms}()},
\code{\link{slackrSetup}()},
\code{\link{slackrUpload}()},
\code{\link{slackrUsers}()},
\code{\link{textSlackr}()}
}
\concept{Deprecated functions}
