% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcamels.r
\name{slackrSetup}
\alias{slackrSetup}
\title{slackrSetup}
\usage{
slackrSetup(...)
}
\arguments{
\item{...}{arguments to pass to slackr_setup}
}
\description{
deprecated
}
\seealso{
\code{\link[slackr]{slackr_setup}}
}
