% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcamels.r
\name{slackrUpload}
\alias{slackrUpload}
\title{slackrUpload}
\usage{
slackrUpload(...)
}
\arguments{
\item{...}{arguments to pass to slackr_upload}
}
\description{
deprecated
}
\seealso{
\code{\link[slackr]{slackr_upload}}
}
