% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr_utils.R
\name{stop_for_status}
\alias{stop_for_status}
\title{Internal function to warn if Slack API call is not ok.}
\usage{
stop_for_status(r)
}
\arguments{
\item{r}{The response from a call to the Slack API}
}
\description{
The function is called for the side effect of warning when the API response has errors, and is a thin wrapper around httr::stop_for_status
}
\keyword{Internal}
