% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\docType{data}
\name{teams}
\alias{teams}
\title{Slack Teams}
\description{
Manage multilple Slack teams
}
\details{
If a json file exists, it will be read on initialization and
add teams listed in the file.
}
\keyword{datasets}
