% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcamels.r
\name{textSlackr}
\alias{textSlackr}
\title{textSlackr}
\usage{
textSlackr(...)
}
\arguments{
\item{...}{arguments to pass to slackr_msg}
}
\description{
deprecated
}
\seealso{
\code{\link[slackr]{slackr_msg}}
}
