% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_slack_api.R
\name{with_pagination}
\alias{with_pagination}
\title{Calls the slack API with pagination using cursors.}
\usage{
with_pagination(fun, extract)
}
\arguments{
\item{fun}{A function that calls the slack API}

\item{extract}{The name of the element to extract from the API reponse}
}
\value{
A \code{tibble}
}
\description{
This loops over \code{fun}, extracts the \code{next_cursor} from the API response, and
injects this into the next loop.  At the completion of each loop, the function \code{\link[=convert_response_to_tibble]{convert_response_to_tibble()}} is run with \code{extract} as and argument. The results are combined with \code{\link[dplyr:bind]{dplyr::bind_rows()}}
}
\seealso{
call_slack_api
}
