% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_slack_api.R
\name{auth_test}
\alias{auth_test}
\title{Checks authentication & identity against the Slack API.}
\usage{
auth_test(token = Sys.getenv("SLACK_TOKEN"))
}
\arguments{
\item{token}{Authentication token bearing required scopes.}
}
\description{
Checks authentication & identity against the Slack API.
}
\examples{
if (Sys.getenv("SLACK_TOKEN") != "") {
  auth_test()
}
}
\references{
https://api.slack.com/methods/auth.test
}
