% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dev_slackr.R
\name{dev_slackr}
\alias{dev.slackr}
\alias{dev_slackr}
\title{Send the graphics contents of the current device to a \code{slack.com} channel}
\usage{
dev_slackr(channels = Sys.getenv("SLACK_CHANNEL"), ...,
  api_token = Sys.getenv("SLACK_API_TOKEN"), file = "plot")

dev.slackr(channels = Sys.getenv("SLACK_CHANNEL"), ...,
  api_token = Sys.getenv("SLACK_API_TOKEN"), file = "plot")
}
\arguments{
\item{channels}{list of channels to post image to}

\item{...}{other arguments passed into png device}

\item{api_token}{the slack.com full API token (chr)}

\item{file}{prefix for filenames (defaults to \code{plot})}
}
\value{
\code{httr} response object from \code{POST} call
}
\description{
\code{dev.slackr} sends the graphics contents of the current device to the specified \code{slack.com} channel.
This requires setting up a full API token (i.e. not a webhook & not OAuth) for this to work.
}
\note{
You can pass in \code{add_user=TRUE} as part of the \code{...} parameters and the Slack API
      will post the message as your logged-in user account (this will override anything set in
      \code{username})
}
\examples{
\dontrun{
slackr_setup()

# base
library(maps)
maps("usa")
dev_slackr("#results", filename='map')

# base
barplot(VADeaths)
dev_slackr("@jayjacobs")
}
}
\author{
Konrad Karczewski [ctb], Bob Rudis [aut]
}
\references{
\url{https://github.com/hrbrmstr/slackr/pull/12/files}
}
\seealso{
\code{\link{slackrSetup}}, \code{\link{save.slackr}}, \code{\link{slackrUpload}}
}

