% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\name{dev_slackr}
\alias{dev_slackr}
\title{Superseded by slackr_dev}
\usage{
dev_slackr(...)
}
\arguments{
\item{...}{arguments to pass to slackr_dev}
}
\description{
deprecated
}
\seealso{
\code{\link[slackr]{slackr_dev}}

Other Deprecated functions: 
\code{\link{dev.slackr}()},
\code{\link{save.slackr}()},
\code{\link{save_slackr}()},
\code{\link{slackrBot}()},
\code{\link{slackrChannels}()},
\code{\link{slackrChtrans}()},
\code{\link{slackrIms}()},
\code{\link{slackrMsg}()},
\code{\link{slackrSetup}()},
\code{\link{slackrUpload}()},
\code{\link{slackrUsers}()},
\code{\link{tex_slackr}()},
\code{\link{textSlackr}()}
}
\concept{Deprecated functions}
