% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ggslackr}
\alias{ggslackr}
\title{Post a ggplot to a \url{slack.com} channel}
\usage{
ggslackr(api_token = Sys.getenv("SLACK_API_TOKEN"),
  channels = Sys.getenv("SLACK_CHANNEL"), ...)
}
\arguments{
\item{api_token}{the slack.com full API token (chr)}

\item{channels}{list of channels to post image to}

\item{...}{other arguments passed into png device}
}
\description{
need to setup a full API token (i.e. not a webhook & not OAuth) for this to work
}

