% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\name{save.slackr}
\alias{save.slackr}
\title{Superseded by save_slackr}
\usage{
save.slackr(...)
}
\arguments{
\item{...}{arguments to pass to save_slackr}
}
\description{
deprecated
}
\seealso{
\code{\link[slackr]{save_slackr}}

Other Deprecated functions: 
\code{\link{dev.slackr}()},
\code{\link{dev_slackr}()},
\code{\link{save_slackr}()},
\code{\link{slackrBot}()},
\code{\link{slackrChannels}()},
\code{\link{slackrChtrans}()},
\code{\link{slackrIms}()},
\code{\link{slackrMsg}()},
\code{\link{slackrSetup}()},
\code{\link{slackrUpload}()},
\code{\link{slackrUsers}()},
\code{\link{tex_slackr}()},
\code{\link{textSlackr}()}
}
\concept{Deprecated functions}
