% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{save.slackr}
\alias{save.slackr}
\title{Save R objects to an RData file on \code{slack.com}}
\usage{
save.slackr(..., channels = "", file = "slackr",
  api_token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{...}{objects to store in the R data file}

\item{channels}{slack.com channels to save to (optional)}

\item{file}{filename (without extension) to use}

\item{api_token}{full API token}
}
\value{
\code{httr} response object from \code{POST} call
}
\description{
\code{save.slackr} enables you upload R objects (as an R data file)
to \code{slack.com} and (optionally) post them to one or more channels
(if \code{channels} is not empty).
}
\seealso{
\code{\link{slackrSetup}}, \code{\link{dev.slackr}}, \code{\link{slackrUpload}}
}

