% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr.R
\name{slackr}
\alias{slackr}
\title{Output R expressions to a Slack channel/user}
\usage{
slackr(
  ...,
  channel = Sys.getenv("SLACK_CHANNEL"),
  username = Sys.getenv("SLACK_USERNAME"),
  icon_emoji = Sys.getenv("SLACK_ICON_EMOJI"),
  token = Sys.getenv("SLACK_TOKEN"),
  thread_ts = NULL,
  reply_broadcast = FALSE
)
}
\arguments{
\item{...}{expressions to be sent to Slack.}

\item{channel}{Channel, private group, or IM channel to send message to. Can be an encoded ID, or a name. See the \href{https://api.slack.com/methods/chat.postMessage#channels}{chat.postMessage endpoint documentation} for details.}

\item{username}{what user should the bot be named as (chr).}

\item{icon_emoji}{what emoji to use (chr) \code{""} will mean use the default.}

\item{token}{Authentication token bearing required scopes.}

\item{thread_ts}{Provide another message's ts value to make this message a reply. Avoid using a reply's ts value; use its parent instead.}

\item{reply_broadcast}{Used in conjunction with thread_ts and indicates whether reply should be made visible to everyone in the channel or conversation. Defaults to FALSE.}
}
\value{
the response (invisibly)
}
\description{
Takes an \code{expr}, evaluates it and sends the output to a Slack
chat destination. Useful for logging, messaging on long compute tasks or
general information sharing.
}
\details{
By default, everything but \code{expr} will be looked for in a "\code{SLACK_}"
environment variable. You can override or just specify these values directly instead,
but it's probably better to call \code{\link[=slackr_setup]{slackr_setup()}} first.
}
\note{
You need a \url{https://www.slack.com} account and will also need to
set up an API token \url{https://api.slack.com/}
}
\examples{
\dontrun{
slackr_setup()
slackr("iris info", head(iris), str(iris))
}
}
\seealso{
\code{\link[=slackr_setup]{slackr_setup()}}, \code{\link[=slackr_bot]{slackr_bot()}}, \code{\link[=slackr_dev]{slackr_dev()}},
\code{\link[=slackr_save]{slackr_save()}}, \code{\link[=slackr_upload]{slackr_upload()}}
}
