% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr.R
\name{slackr}
\alias{slackr}
\title{Output R expressions to a Slack channel/user}
\usage{
slackr(
  ...,
  channel = Sys.getenv("SLACK_CHANNEL"),
  username = Sys.getenv("SLACK_USERNAME"),
  icon_emoji = Sys.getenv("SLACK_ICON_EMOJI"),
  bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN")
)
}
\arguments{
\item{...}{expressions to be sent to Slack}

\item{channel}{which channel to post the message to (chr)}

\item{username}{what user should the bot be named as (chr)}

\item{icon_emoji}{what emoji to use (chr) \code{""} will mean use the default}

\item{bot_user_oauth_token}{Slack bot user OAuth token}
}
\value{
the response (invisibly)
}
\description{
Takes an \code{expr}, evaluates it and sends the output to a Slack
chat destination. Useful for logging, messaging on long compute tasks or
general information sharing.
}
\details{
By default, everything but \code{expr} will be looked for in a "\code{SLACK_}"
environment variable. You can override or just specify these values directly instead,
but it's probably better to call \code{\link[=slackr_setup]{slackr_setup()}} first.
}
\note{
You need a \url{https://www.slack.com} account and will also need to
setup an API token \url{https://api.slack.com/}
Also, you can pass in \code{as_user=TRUE}, the default, as part of the \code{...}
parameters and the Slack API will post the message as your logged-in
user account (this will override anything set in \code{username}).
Passing \code{as_user=FALSE}, results in the Slack API posting the
message as set in \code{username}
}
\examples{
\dontrun{
slackr_setup()
slackr("iris info", head(iris), str(iris))
}
}
\seealso{
\code{\link[=slackr_setup]{slackr_setup()}}, \code{\link[=slackr_bot]{slackr_bot()}}, \code{\link[=dev_slackr]{dev_slackr()}},
\code{\link[=save_slackr]{save_slackr()}}, \code{\link[=slackr_upload]{slackr_upload()}}
}
