% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr_utils.R, R/zcamels.r
\name{slackr_channels}
\alias{slackr_channels}
\alias{slackrChannels}
\title{Get a data frame of Slack channels}
\usage{
slackr_channels(api_token = Sys.getenv("SLACK_API_TOKEN"))

slackrChannels(api_token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{api_token}{the Slack full API token (chr)}
}
\value{
data.table of channels
}
\description{
Get a data frame of Slack channels
}
