% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr_utils.R, R/zcamels.r
\name{slackr_channels}
\alias{slackr_channels}
\alias{slackrChannels}
\title{Get a data frame of Slack channels}
\usage{
slackr_channels(
  bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN")
)

slackrChannels(bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN"))
}
\arguments{
\item{bot_user_oauth_token}{the Slack bot OAuth token (chr)}
}
\value{
data.table of channels
}
\description{
Get a data frame of Slack channels
}
