% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr_utils.R
\name{slackr_channels}
\alias{slackr_channels}
\title{Get a data frame of Slack channels}
\usage{
slackr_channels(
  token = Sys.getenv("SLACK_TOKEN"),
  bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN")
)
}
\arguments{
\item{token}{A Slack token (either a user token or a bot user token)}

\item{bot_user_oauth_token}{Deprecated. A Slack bot user OAuth token}
}
\value{
data.table of channels
}
\description{
Get a data frame of Slack channels
}
