% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr_utils.R
\name{slackr_channels}
\alias{slackr_channels}
\title{Get a data frame of Slack channels}
\usage{
slackr_channels(token = Sys.getenv("SLACK_TOKEN"))
}
\arguments{
\item{token}{Authentication token bearing required scopes.}
}
\value{
data.table of channels
}
\description{
Get a data frame of Slack channels
}
