% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr_utils.R, R/zcamels.r
\name{slackr_chtrans}
\alias{slackr_chtrans}
\alias{slackrChtrans}
\alias{slackrChTrans}
\title{Translate vector of channel names to channel ID's for API}
\usage{
slackr_chtrans(channels, api_token = Sys.getenv("SLACK_API_TOKEN"),
  census = getGlobalIfMissing("slackr_census"))

slackrChtrans(channels, api_token = Sys.getenv("SLACK_API_TOKEN"),
  census = getGlobalIfMissing("slackr_census"))

slackrChTrans(channels, api_token = Sys.getenv("SLACK_API_TOKEN"),
  census = getGlobalIfMissing("slackr_census"))
}
\arguments{
\item{channels}{vector of channel names to parse}

\item{api_token}{the Slack full API token (chr)}

\item{census}{An object in the format created by \code{\link{runcensus}}: if not supplied,
it looks for an object named \code{slackr_census} in your global environment. If none is found,
it performs \code{\link{runcensus}}.}
}
\value{
character vector - original channel list with \code{#} or
         \code{@} channels replaced with ID's.
}
\description{
Given a vector of one or more channel names, it will retrieve list of
active channels and try to replace channels that begin with "\code{#}" or "\code{@}"
with the channel ID for that channel. Also incorporates groups.
}
\author{
Quinn Weber [ctb], Bob Rudis [aut]
}
