% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr_utils.R, R/zcamels.r
\name{slackr_chtrans}
\alias{slackrChTrans}
\alias{slackrChtrans}
\alias{slackr_chtrans}
\title{Translate vector of channel names to channel ID's for API}
\usage{
slackr_chtrans(channels, api_token = Sys.getenv("SLACK_API_TOKEN"))

slackrChtrans(channels, api_token = Sys.getenv("SLACK_API_TOKEN"))

slackrChTrans(channels, api_token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{channels}{vector of channel names to parse}

\item{api_token}{the slack.com full API token (chr)}
}
\value{
character vector - original channel list with \code{#} or
         \code{@} channels replaced with ID's.
}
\description{
Given a vector of one or more channel names, it will retrieve list of
active channels and try to replace channels that begin with "\code{#}" or "\code{@}"
with the channel ID for that channel. Also incorporates groups.
}
\author{
Quinn Weber [ctb], Bob Rudis [aut]
}

