% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr_utils.R
\name{slackr_chtrans}
\alias{slackr_chtrans}
\title{Translate vector of channel names to channel IDs for API}
\usage{
slackr_chtrans(channels, token = Sys.getenv("SLACK_TOKEN"))
}
\arguments{
\item{channels}{Comma-separated list of channel names or IDs where the file will be shared.}

\item{token}{Authentication token bearing required scopes.}
}
\value{
character vector - original channel list with \verb{#} or
\code{@} channels replaced with ID's.
}
\description{
Given a vector of one or more channel names, retrieve list of
active channels and try to replace channels that begin with "\verb{#}" or "\code{@}"
with the channel ID for that channel.
}
\author{
Quinn Weber (ctb), Bob Rudis (aut)
}
