% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr_utils.R, R/zcamels.r
\name{slackr_chtrans}
\alias{slackr_chtrans}
\alias{slackrChtrans}
\alias{slackrChTrans}
\title{Translate vector of channel names to channel ID's for API}
\usage{
slackr_chtrans(
  channels,
  bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN")
)

slackrChtrans(
  channels,
  bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN")
)

slackrChTrans(
  channels,
  bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN")
)
}
\arguments{
\item{channels}{vector of channel names to parse}

\item{bot_user_oauth_token}{the Slack bot user OAuth token (chr)}
}
\value{
character vector - original channel list with \code{#} or
         \code{@} channels replaced with ID's.
}
\description{
Given a vector of one or more channel names, it will retrieve list of
active channels and try to replace channels that begin with "\code{#}" or "\code{@}"
with the channel ID for that channel.
}
\author{
Quinn Weber [ctb], Bob Rudis [aut]
}
