% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr_delete.R
\name{slackr_delete}
\alias{slackr_delete}
\title{slackr_delete}
\usage{
slackr_delete(
  count,
  channel = Sys.getenv("SLACK_CHANNEL"),
  token = Sys.getenv("SLACK_TOKEN")
)
}
\arguments{
\item{count}{The number of messages to delete.}

\item{channel}{Channel, private group, or IM channel to delete messages from. Can be an encoded ID, or a name. See the \href{https://api.slack.com/methods/chat.postMessage#channels}{chat.postMessage endpoint documentation} for details.}

\item{token}{Authentication token bearing required scopes.}
}
\description{
Deletes the specified number of messages from the channel
}
