% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr_utils.R, R/zcamels.r
\name{slackr_groups}
\alias{slackrGroups}
\alias{slackr_groups}
\title{Get a data frame of Slack groups}
\usage{
slackr_groups(api_token = Sys.getenv("SLACK_API_TOKEN"))

slackrGroups(api_token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{api_token}{the Slackfull API token (chr)}
}
\value{
\code{data.frame} of channels
}
\description{
Get a data frame of Slack groups
}

