% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr_history.R
\name{slackr_history}
\alias{slackr_history}
\title{slackr_history}
\usage{
slackr_history(
  bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN"),
  channel = Sys.getenv("SLACK_CHANNEL"),
  posted_from_time = 0L,
  posted_to_time = as.numeric(Sys.time()),
  message_count = 100L
)
}
\arguments{
\item{bot_user_oauth_token}{the Slack bot user OAuth token}

\item{channel}{the channel to get history from}

\item{posted_from_time}{the first post time to consider (default: time = 0)}

\item{posted_to_time}{the last post time to consider (default: current time)}

\item{message_count}{the number of messages to retain}
}
\description{
Returns a dataframe of post history in a channel
}
