% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slackr_utils.R
\name{slackr_ims}
\alias{slackrIms}
\alias{slackr_ims}
\title{Get a data frame of slack.com IM ids}
\usage{
slackr_ims(api_token = Sys.getenv("SLACK_API_TOKEN"))

slackrIms(api_token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{api_token}{the slack.com full API token (chr)}
}
\value{
data.table of im ids and user names
}
\description{
need to setup a full API token (i.e. not a webhook & not OAuth) for this to work
}
\author{
Quinn Weber [aut], Bob Rudis [ctb]
}
\references{
\url{https://github.com/hrbrmstr/slackr/pull/13}
}

