% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr_utils.R
\name{slackr_ims}
\alias{slackr_ims}
\title{Get a data frame of Slack IM ids}
\usage{
slackr_ims(
  token = Sys.getenv("SLACK_TOKEN"),
  bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN")
)
}
\arguments{
\item{token}{A Slack token (either a user token or a bot user token)}

\item{bot_user_oauth_token}{Deprecated. A Slack bot user OAuth token}
}
\value{
\code{data.frame} of im ids and user names
}
\description{
Get a data frame of Slack IM ids
}
\references{
\url{https://github.com/mrkaye97/slackr/pull/13}
}
\author{
Quinn Weber (aut), Bob Rudis (ctb)
}
