% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr_utils.R, R/zcamels.r
\name{slackr_ims}
\alias{slackrIms}
\alias{slackr_ims}
\title{Get a data frame of Slack IM ids}
\usage{
slackr_ims(api_token = Sys.getenv("SLACK_API_TOKEN"))

slackrIms(api_token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{api_token}{the Slack full API token (chr)}
}
\value{
\code{data.frame} of im ids and user names
}
\description{
Get a data frame of Slack IM ids
}
\author{
Quinn Weber [aut], Bob Rudis [ctb]
}
\references{
\url{https://github.com/hrbrmstr/slackr/pull/13}
}

