% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr_utils.R, R/zcamels.r
\name{slackr_users}
\alias{slackrUsers}
\alias{slackr_users}
\title{Get a data frame of Slack users}
\usage{
slackr_users(api_token = Sys.getenv("SLACK_API_TOKEN"))

slackrUsers(api_token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{api_token}{the Slack full API token (chr)}
}
\value{
\code{data.frame} of users
}
\description{
Get a data frame of Slack users
}

