% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr_utils.R
\name{slackr_users}
\alias{slackr_users}
\title{Get a data frame of Slack users}
\usage{
slackr_users(bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN"))
}
\arguments{
\item{bot_user_oauth_token}{the Slack bot OAuth token (chr)}
}
\value{
\code{data.frame} of users
}
\description{
Get a data frame of Slack users
}
