% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr_utils.R
\name{slackr_users}
\alias{slackr_users}
\title{Get a data frame of Slack users}
\usage{
slackr_users(token = Sys.getenv("SLACK_TOKEN"))
}
\arguments{
\item{token}{Authentication token bearing required scopes.}
}
\value{
\code{data.frame} of users
}
\description{
Get a data frame of Slack users
}
