% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\name{slackrBot}
\alias{slackrBot}
\title{Superseded by slackr_bot}
\usage{
slackrBot(...)
}
\arguments{
\item{...}{arguments to pass to slackr_bot}
}
\description{
deprecated
}
\seealso{
\code{\link[slackr]{slackr_bot}}

Other Deprecated functions: 
\code{\link{dev.slackr}()},
\code{\link{dev_slackr}()},
\code{\link{save.slackr}()},
\code{\link{slackrChannels}()},
\code{\link{slackrChtrans}()},
\code{\link{slackrIms}()},
\code{\link{slackrMsg}()},
\code{\link{slackrSetup}()},
\code{\link{slackrUpload}()},
\code{\link{slackrUsers}()},
\code{\link{textSlackr}()}
}
\concept{Deprecated functions}
