% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{slackrChannels}
\alias{slackrChannels}
\title{Get a data frame of slack.com channels}
\usage{
slackrChannels(api_token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{api_token}{the slack.com full API token (chr)}
}
\value{
data.table of channels
}
\description{
need to setup a full API token (i.e. not a webhook & not OAuth) for this to work
}
\note{
Renamed from \code{slackr_channels}
}

