% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/slackr.R
\name{slackrChTrans}
\alias{slackrChTrans}
\title{Translate vector of channel names to channel ID's for API}
\usage{
slackrChTrans(channels, api_token = Sys.getenv("SLACK_API_TOKEN"),
  ims = FALSE)
}
\arguments{
\item{channels}{vector of channel names to parse}

\item{api_token}{the slack.com full API token (chr)}

\item{ims}{Should the IM ids be used instead of the user ids?
Defaults to FALSE, which is how the function behaved through version 1.2.3}
}
\value{
character vector - original channel list with \code{#} or \code{@} channels replaced with ID's.
}
\description{
Given a vector of one or more channel names, it will retrieve list of
active channels and try to replace channels that begin with "\code{#}" or "\code{@}"
with the channel ID for that channel. Also incorporates groups.
}
\note{
Renamed from \code{slackr_chtrans}
}

