% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcamels.r
\name{slackrChTrans}
\alias{slackrChTrans}
\title{slackrChTrans}
\usage{
slackrChTrans(...)
}
\arguments{
\item{...}{arguments to pass to slackr_chtrans}
}
\description{
deprecated
}
\seealso{
\code{\link[slackr]{slackr_chtrans}}
}
