% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{slackrChTrans}
\alias{slackrChTrans}
\title{Translate vector of channel names to channel ID's for API}
\usage{
slackrChTrans(channels, api_token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{channels}{vector of channel names to parse}

\item{api_token}{the slack.com full API token (chr)}
}
\value{
character vector - original channel list with \code{#} or \code{@} channels replaced with ID's.
}
\description{
Given a vector of one or more channel names, it will retrieve list of
active channels and try to replace channels that begin with "\code{#}" or "\code{@}"
with the channel ID for that channel. Also incorporates groups.
}
\note{
Renamed from \code{slackr_chtrans}
}

