% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\name{slackrChtrans}
\alias{slackrChtrans}
\alias{slackrChTrans}
\title{Superseded by slackr_chtrans}
\usage{
slackrChtrans(...)

slackrChTrans(...)
}
\arguments{
\item{...}{arguments to pass to slackr_chtrans}
}
\description{
deprecated

deprecated
}
\seealso{
\code{\link[slackr]{slackr_chtrans}}

\code{\link[slackr]{slackr_chtrans}}

Other Deprecated functions: 
\code{\link{dev.slackr}()},
\code{\link{dev_slackr}()},
\code{\link{save.slackr}()},
\code{\link{save_slackr}()},
\code{\link{slackrBot}()},
\code{\link{slackrChannels}()},
\code{\link{slackrIms}()},
\code{\link{slackrMsg}()},
\code{\link{slackrSetup}()},
\code{\link{slackrUpload}()},
\code{\link{slackrUsers}()},
\code{\link{tex_slackr}()},
\code{\link{textSlackr}()}

Other Deprecated functions: 
\code{\link{dev.slackr}()},
\code{\link{dev_slackr}()},
\code{\link{save.slackr}()},
\code{\link{save_slackr}()},
\code{\link{slackrBot}()},
\code{\link{slackrChannels}()},
\code{\link{slackrIms}()},
\code{\link{slackrMsg}()},
\code{\link{slackrSetup}()},
\code{\link{slackrUpload}()},
\code{\link{slackrUsers}()},
\code{\link{tex_slackr}()},
\code{\link{textSlackr}()}
}
\concept{Deprecated functions}
