% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\name{slackrIms}
\alias{slackrIms}
\title{Superseded by slackr_ims}
\usage{
slackrIms(...)
}
\arguments{
\item{...}{arguments to pass to slackr_ims}
}
\description{
deprecated
}
\seealso{
\code{\link[slackr]{slackr_ims}}

Other Deprecated functions: 
\code{\link{dev.slackr}()},
\code{\link{dev_slackr}()},
\code{\link{save.slackr}()},
\code{\link{save_slackr}()},
\code{\link{slackrBot}()},
\code{\link{slackrChannels}()},
\code{\link{slackrChtrans}()},
\code{\link{slackrMsg}()},
\code{\link{slackrSetup}()},
\code{\link{slackrUpload}()},
\code{\link{slackrUsers}()},
\code{\link{tex_slackr}()},
\code{\link{textSlackr}()}
}
\concept{Deprecated functions}
