% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{slackrSetup}
\alias{slackrSetup}
\title{Setup environment variables for Slack.com API}
\usage{
slackrSetup(channel = "#general", username = "slackr", icon_emoji = "",
  token = "", url_prefix = "")
}
\arguments{
\item{channel}{which default channel to send the output to (chr) defaults to \code{#general}}

\item{username}{which username will output appear from (chr) defaults to \code{slackr}}

\item{icon_emoji}{which emoji picture to use (chr) defaults to none}

\item{token}{the slack.com API token string (chr) defaults to none}

\item{url_prefix}{the slack.com URL prefix to use (chr) defaults to none}
}
\description{
Initialize all the environment variables \code{slackr()} will need to use to
work properly.
}
\details{
By default, \code{slackr} will use the \code{#general} room and a username
of \code{slackr()} with no emoji and the default slack.com API prefix URL. You
still need to provide the webhook API token in \code{token} for anyting to work.
Failure to call this function before calling \code{slackr()} will result in a
message to do so.
}
\examples{
\dontrun{
slackrSetup(channel="#code", url_prefix="http://myslack.slack.com/services/hooks/incoming-webhook?")
}
}

