% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\name{slackrSetup}
\alias{slackrSetup}
\title{Superseded by slackr_setup}
\usage{
slackrSetup(...)
}
\arguments{
\item{...}{arguments to pass to slackr_setup}
}
\description{
deprecated
}
\seealso{
\code{\link[slackr]{slackr_setup}}

Other Deprecated functions: 
\code{\link{dev.slackr}()},
\code{\link{dev_slackr}()},
\code{\link{save.slackr}()},
\code{\link{slackrBot}()},
\code{\link{slackrChannels}()},
\code{\link{slackrChtrans}()},
\code{\link{slackrIms}()},
\code{\link{slackrMsg}()},
\code{\link{slackrUpload}()},
\code{\link{slackrUsers}()},
\code{\link{textSlackr}()}
}
\concept{Deprecated functions}
