% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\name{slackrUpload}
\alias{slackrUpload}
\title{Superseded by slackr_upload}
\usage{
slackrUpload(...)
}
\arguments{
\item{...}{arguments to pass to slackr_upload}
}
\description{
deprecated
}
\seealso{
\code{\link[slackr]{slackr_upload}}

Other Deprecated functions: 
\code{\link{dev.slackr}()},
\code{\link{dev_slackr}()},
\code{\link{save.slackr}()},
\code{\link{save_slackr}()},
\code{\link{slackrBot}()},
\code{\link{slackrChannels}()},
\code{\link{slackrChtrans}()},
\code{\link{slackrIms}()},
\code{\link{slackrMsg}()},
\code{\link{slackrSetup}()},
\code{\link{slackrUsers}()},
\code{\link{tex_slackr}()},
\code{\link{textSlackr}()}
}
\concept{Deprecated functions}
