% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\name{slackrUsers}
\alias{slackrUsers}
\title{Superseded by slackr_users}
\usage{
slackrUsers(...)
}
\arguments{
\item{...}{arguments to pass to slackr_users}
}
\description{
deprecated
}
\seealso{
\code{\link[slackr]{slackr_users}}

Other Deprecated functions: 
\code{\link{dev.slackr}()},
\code{\link{dev_slackr}()},
\code{\link{save.slackr}()},
\code{\link{slackrBot}()},
\code{\link{slackrChannels}()},
\code{\link{slackrChtrans}()},
\code{\link{slackrIms}()},
\code{\link{slackrMsg}()},
\code{\link{slackrSetup}()},
\code{\link{slackrUpload}()},
\code{\link{textSlackr}()}
}
\concept{Deprecated functions}
