% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\name{textSlackr}
\alias{textSlackr}
\title{Superseded by text_slackr}
\usage{
textSlackr(...)
}
\arguments{
\item{...}{arguments to pass to text_slackr}
}
\description{
deprecated
}
\seealso{
\code{\link[slackr]{slackr_msg}}

Other Deprecated functions: 
\code{\link{dev.slackr}()},
\code{\link{dev_slackr}()},
\code{\link{save.slackr}()},
\code{\link{save_slackr}()},
\code{\link{slackrBot}()},
\code{\link{slackrChannels}()},
\code{\link{slackrChtrans}()},
\code{\link{slackrIms}()},
\code{\link{slackrMsg}()},
\code{\link{slackrSetup}()},
\code{\link{slackrUpload}()},
\code{\link{slackrUsers}()},
\code{\link{tex_slackr}()}
}
\concept{Deprecated functions}
