% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{warn_for_args}
\alias{warn_for_args}
\title{Check for token-parameter mismatches}
\usage{
warn_for_args(token, ...)
}
\arguments{
\item{token}{a token}

\item{...}{Additiional arguments passed to the function called}
}
\value{
No return value. Called for side effects
}
\description{
Check for token-parameter mismatches
}
